refactor(tui): split input.rs into parse.rs and display.rs (#583)
Extract stateless message parsing utilities (parse_dm_input,
build_payload, seed_online_users_from_who, parse_status_broadcast,
parse_kick_broadcast, parse_subscription_broadcast, apply_backslash_enter,
normalize_paste) to tui/parse.rs.

Extract pure text/cursor calculation functions (wrap_input_display,
cursor_display_pos, byte_offset_at_display_pos, find_at_context,
prev_word_start, next_word_end) to tui/display.rs.

input.rs retains InputState, Action, handle_key dispatcher, per-key
handlers, and picker sync functions (~497 prod lines, under 500 target).
Tests move with their functions. No API changes — all functions
re-exported at the same paths.

Closes #583
