+++ crates/room-cli/src/tui/input.rs
@@ -1,8 +1,12 @@
-use std::collections::HashMap;
-
-use unicode_width::UnicodeWidthChar;
-
-use room_protocol::SubscriptionTier;
+pub(super) use super::display::{
+    byte_offset_at_display_pos, cursor_display_pos, find_at_context, next_word_end,
+    prev_word_start, wrap_input_display,
+};
+use super::parse::parse_dm_input;
+pub(super) use super::parse::{
+    apply_backslash_enter, build_payload, normalize_paste, parse_kick_broadcast,
+    parse_status_broadcast, parse_subscription_broadcast, seed_online_users_from_who,
+};
 
 use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
 
@@ -368,139 +372,6 @@ fn handle_right(state: &mut InputState, online_users: &[String]) {
     sync_mention_after_cursor_move(state, online_users);
 }
 
-/// Wrap input text for display in the input box.
-///
-/// Splits on `\n` (explicit newlines from Shift+Enter), then wraps each
-/// logical line character-by-character at `width` display columns using
-/// Unicode display widths. Returns the flat list of display rows.
-///
-/// If `width` is 0, returns each logical line unsplit.
-pub(super) fn wrap_input_display(input: &str, width: usize) -> Vec<String> {
-    let mut rows: Vec<String> = Vec::new();
-    for logical_line in input.split('\n') {
-        if width == 0 {
-            rows.push(logical_line.to_string());
-            continue;
-        }
-        let mut current = String::new();
-        let mut col: usize = 0;
-        for ch in logical_line.chars() {
-            let w = ch.width().unwrap_or(0);
-            // Wrap before adding if the character would overflow and we have content.
-            if col + w > width && col > 0 {
-                rows.push(std::mem::take(&mut current));
-                col = 0;
-            }
-            current.push(ch);
-            col += w;
-        }
-        rows.push(current);
-    }
-    // Always at least one row.
-    if rows.is_empty() {
-        rows.push(String::new());
-    }
-    rows
-}
-
-/// Return the `(display_row, display_col)` of `cursor_pos` (a byte index into
-/// `input`) after applying the same wrapping logic as [`wrap_input_display`].
-pub(super) fn cursor_display_pos(input: &str, cursor_pos: usize, width: usize) -> (usize, usize) {
-    let mut row: usize = 0;
-    let mut col: usize = 0;
-    for (i, ch) in input.char_indices() {
-        if i >= cursor_pos {
-            break;
-        }
-        if ch == '\n' {
-            row += 1;
-            col = 0;
-        } else {
-            let w = ch.width().unwrap_or(0);
-            if width > 0 && col + w > width && col > 0 {
-                row += 1;
-                col = 0;
-            }
-            col += w;
-        }
-    }
-    // If col == width and there is remaining content that is not a newline,
-    // the next character would start on a new row — advance the cursor there
-    // so it doesn't render past the right border of the input box.
-    if width > 0
-        && col == width
-        && cursor_pos < input.len()
-        && !input[cursor_pos..].starts_with('\n')
-    {
-        row += 1;
-        col = 0;
-    }
-    (row, col)
-}
-
-/// Given a target `(display_row, display_col)` in the wrapped display of `input`,
-/// return the byte offset into `input` that corresponds to that position.
-///
-/// Uses the same wrapping logic as [`wrap_input_display`] and [`cursor_display_pos`].
-/// If `target_col` is beyond the end of `target_row`, the returned offset is
-/// clamped to the last character position in that row (before any soft-wrap or newline).
-pub(super) fn byte_offset_at_display_pos(
-    input: &str,
-    target_row: usize,
-    target_col: usize,
-    width: usize,
-) -> usize {
-    let mut row: usize = 0;
-    let mut col: usize = 0;
-
-    for (i, ch) in input.char_indices() {
-        if row == target_row {
-            if col >= target_col {
-                return i;
-            }
-            // About to advance past the target row — clamp before the newline.
-            if ch == '\n' {
-                return i;
-            }
-        }
-
-        if ch == '\n' {
-            row += 1;
-            col = 0;
-        } else {
-            let w = ch.width().unwrap_or(0);
-            if width > 0 && col + w > width && col > 0 {
-                // This character starts a new soft-wrapped row.
-                if row == target_row {
-                    // target_col was beyond the end of the soft-wrapped row.
-                    return i;
-                }
-                row += 1;
-                col = 0;
-                if row == target_row && col >= target_col {
-                    return i;
-                }
-            }
-            col += w;
-        }
-    }
-
-    input.len()
-}
-
-/// If the cursor is currently within a `@mention` context — i.e. there is an
-/// `@` before `cursor_pos` with no space or newline between them — return the
-/// byte index of the `@` and the query string typed after it.
-pub(super) fn find_at_context(input: &str, cursor_pos: usize) -> Option<(usize, &str)> {
-    let before = &input[..cursor_pos];
-    let at_pos = before.rfind('@')?;
-    let query = &before[at_pos + 1..];
-    if query.contains([' ', '\n']) {
-        return None;
-    }
-    Some((at_pos, query))
-}
-
 /// Re-evaluate the mention picker after the cursor moved.
 ///
 /// If the picker is active, checks the `@` context at the new cursor position.
@@ -620,206 +491,7 @@ fn sync_palette_after_delete(state: &mut InputState) {
     }
 }
 
-/// If the char immediately before `cursor_pos` in `buf` is `\`, removes it
-/// and inserts `\n`, returning the new cursor position. Returns `None` if the
-/// precondition is not met (no preceding backslash or cursor at start).
-///
-/// This mirrors the backslash+Enter key binding in the TUI event loop.
-pub(super) fn apply_backslash_enter(buf: &mut String, cursor_pos: usize) -> Option<usize> {
-    if cursor_pos > 0 && buf[..cursor_pos].ends_with('\\') {
-        let bs_pos = cursor_pos - 1; // '\\' is ASCII (1 byte)
-        buf.remove(bs_pos);
-        buf.insert(bs_pos, '\n');
-        Some(bs_pos + 1)
-    } else {
-        None
-    }
-}
-
-/// Normalize pasted text: convert `\r\n` to `\n`, then stray `\r` to `\n`.
-///
-/// Called on `Event::Paste` to ensure consistent newline handling regardless
-/// of the clipboard's line-ending convention (Windows, old Mac, Unix).
-pub(super) fn normalize_paste(text: &str) -> String {
-    text.replace("\r\n", "\n").replace('\r', "\n")
-}
-
-/// Parse `/dm <user> <message>` input into a `DmRoom` action.
-///
-/// Returns `Some(Action::DmRoom { .. })` when the input is a valid `/dm`
-/// command with both a target user and message content. Returns `None` for
-/// incomplete input (missing user or message) — those fall through to
-/// `build_payload` for backwards-compatible intra-room DM handling.
-fn parse_dm_input(input: &str) -> Option<Action> {
-    let rest = input.strip_prefix("/dm ")?;
-    let mut parts = rest.splitn(2, ' ');
-    let target_user = parts.next().filter(|s| !s.is_empty())?;
-    let content = parts.next().filter(|s| !s.is_empty())?;
-    Some(Action::DmRoom {
-        target_user: target_user.to_owned(),
-        content: content.to_owned(),
-    })
-}
-
-/// Convert TUI input to a JSON envelope for the broker.
-pub(super) fn build_payload(input: &str) -> String {
-    // `/dm <user> <message>` — preserve spaces in the message body.
-    // NOTE: This branch is now only reached when `parse_dm_input` returns
-    // None (e.g. `/dm user` with no message body).
-    if let Some(rest) = input.strip_prefix("/dm ") {
-        let mut parts = rest.splitn(2, ' ');
-        let to = parts.next().unwrap_or("").to_owned();
-        let content = parts.next().unwrap_or("").to_owned();
-        return serde_json::json!({ "type": "dm", "to": to, "content": content }).to_string();
-    }
-    if let Some(rest) = input.strip_prefix('/') {
-        let mut parts = rest.splitn(2, ' ');
-        let cmd = parts.next().unwrap_or("").to_owned();
-        let params: Vec<String> = parts
-            .next()
-            .unwrap_or("")
-            .split_whitespace()
-            .map(str::to_owned)
-            .collect();
-        serde_json::json!({ "type": "command", "cmd": cmd, "params": params }).to_string()
-    } else {
-        serde_json::json!({ "type": "message", "content": input }).to_string()
-    }
-}
-
-/// Seed `online_users` and `user_statuses` from the broker's `/who` response content.
-///
-/// The broker sends `"online — alice, bob: away, charlie"` (or `"no users online"`).
-/// Each entry is either a bare username or `username: status`.
-/// Merges into the existing list without removing users added by Join events.
-pub(super) fn seed_online_users_from_who(
-    content: &str,
-    online_users: &mut Vec<String>,
-    user_statuses: &mut HashMap<String, String>,
-) {
-    if let Some(rest) = content.strip_prefix("online \u{2014} ") {
-        for entry in rest.split(", ") {
-            let (username, status) = match entry.split_once(": ") {
-                Some((u, s)) => (u.trim().to_owned(), s.trim().to_owned()),
-                None => (entry.trim().to_owned(), String::new()),
-            };
-            if !username.is_empty() {
-                if !online_users.contains(&username) {
-                    online_users.push(username.clone());
-                }
-                user_statuses.insert(username, status);
-            }
-        }
-    }
-}
-
-/// Parse a `/set_status` system broadcast into `(username, status)`.
-///
-/// The broker broadcasts either:
-/// - `"alice set status: busy"` → `Some(("alice", "busy"))`
-/// - `"alice cleared their status"` → `Some(("alice", ""))`
-pub(super) fn parse_status_broadcast(content: &str) -> Option<(String, String)> {
-    if let Some(rest) = content.strip_suffix(" cleared their status") {
-        return Some((rest.to_owned(), String::new()));
-    }
-    if let Some((name, status)) = content.split_once(" set status: ") {
-        return Some((name.to_owned(), status.to_owned()));
-    }
-    None
-}
-
-/// Parse a `/kick` system broadcast into the kicked username.
-///
-/// The broker broadcasts `"alice kicked bob (token invalidated)"`.
-/// Returns `Some("bob")` — the target user who was kicked.
-pub(super) fn parse_kick_broadcast(content: &str) -> Option<&str> {
-    let rest = content.strip_suffix(" (token invalidated)")?;
-    let (_issuer, target) = rest.split_once(" kicked ")?;
-    if target.is_empty() {
-        return None;
-    }
-    Some(target)
-}
-
-/// Parse a `/subscribe` system broadcast into `(username, SubscriptionTier)`.
-///
-/// The broker broadcasts `"alice subscribed to room-dev (tier: full)"`.
-/// Returns `Some(("alice", SubscriptionTier::Full))`.
-pub(super) fn parse_subscription_broadcast(content: &str) -> Option<(String, SubscriptionTier)> {
-    // Strip trailing ")"
-    let rest = content.strip_suffix(')')?;
-    // Split on " (tier: "
-    let (name_room, tier_str) = rest.split_once(" (tier: ")?;
-    // Split on " subscribed to "
-    let (name, _room) = name_room.split_once(" subscribed to ")?;
-    if name.is_empty() {
-        return None;
-    }
-    let tier = tier_str.parse().ok()?;
-    Some((name.to_owned(), tier))
-}
-
-/// Move the cursor to the start of the previous word.
-///
-/// "Word" is a maximal run of non-whitespace characters. Starting from
-/// `cursor_pos`, the function first skips any trailing whitespace going
-/// backwards (phase 1), then skips the preceding non-whitespace run (phase 2),
-/// landing at the first byte of that word.
-///
-/// Returns 0 if there is no previous word.
-fn prev_word_start(input: &str, cursor_pos: usize) -> usize {
-    let before = &input[..cursor_pos];
-    let mut it = before.char_indices().rev();
-    // Phase 1: skip trailing whitespace going backward.
-    loop {
-        match it.next() {
-            None => return 0,
-            Some((_, c)) if c.is_whitespace() => continue,
-            Some((i, _)) => {
-                // Found the first non-whitespace; now skip the whole word.
-                // `i` is the byte index of the last char of the preceding word.
-                let mut word_start = i;
-                for (j, c) in it {
-                    if c.is_whitespace() {
-                        // `j` is the byte index of the whitespace char before the word.
-                        return j + c.len_utf8();
-                    }
-                    word_start = j;
-                }
-                return word_start;
-            }
-        }
-    }
-}
-
-/// Move the cursor to one past the end of the next word.
-///
-/// "Word" is a maximal run of non-whitespace characters. Starting from
-/// `cursor_pos`, the function first skips any leading whitespace going forward
-/// (phase 1), then skips the next non-whitespace run (phase 2), landing at the
-/// byte just after the last character of that word.
-///
-/// Returns `input.len()` if there is no next word.
-fn next_word_end(input: &str, cursor_pos: usize) -> usize {
-    let after = &input[cursor_pos..];
-    let mut it = after.char_indices().peekable();
-    // Phase 1: skip leading whitespace.
-    loop {
-        match it.next() {
-            None => return input.len(),
-            Some((_, c)) if c.is_whitespace() => continue,
-            Some(_) => break,
-        }
-    }
-    // Phase 2: skip non-whitespace.
-    loop {
-        match it.next() {
-            None => return input.len(),
-            Some((i, c)) if c.is_whitespace() => return cursor_pos + i,
-            _ => continue,
-        }
-    }
-}
+// Parse/display functions extracted to tui/parse.rs and tui/display.rs.
 
 // ── Tests ─────────────────────────────────────────────────────────────────────
 
@@ -827,479 +499,7 @@ fn next_word_end(input: &str, cursor_pos: usize) -> usize {
 mod tests {
     use super::*;
 
-    // ── seed_online_users_from_who tests ─────────────────────────────────────
-
-    #[test]
-    fn seed_who_populates_users() {
-        let mut users = Vec::new();
-        let mut statuses = HashMap::new();
-        seed_online_users_from_who(
-            "online \u{2014} alice, bob, charlie",
-            &mut users,
-            &mut statuses,
-        );
-        assert_eq!(users, ["alice", "bob", "charlie"]);
-        assert_eq!(statuses.get("alice").unwrap(), "");
-    }
-
-    #[test]
-    fn seed_who_extracts_statuses() {
-        let mut users = Vec::new();
-        let mut statuses = HashMap::new();
-        seed_online_users_from_who(
-            "online \u{2014} alice: away, bob: coding, charlie",
-            &mut users,
-            &mut statuses,
-        );
-        assert_eq!(users, ["alice", "bob", "charlie"]);
-        assert_eq!(statuses.get("alice").unwrap(), "away");
-        assert_eq!(statuses.get("bob").unwrap(), "coding");
-        assert_eq!(statuses.get("charlie").unwrap(), "");
-    }
-
-    #[test]
-    fn seed_who_no_users_online_is_noop() {
-        let mut users = Vec::new();
-        let mut statuses = HashMap::new();
-        seed_online_users_from_who("no users online", &mut users, &mut statuses);
-        assert!(users.is_empty());
-        assert!(statuses.is_empty());
-    }
-
-    #[test]
-    fn seed_who_does_not_duplicate_existing_users() {
-        let mut users = vec!["alice".to_owned()];
-        let mut statuses = HashMap::new();
-        seed_online_users_from_who("online \u{2014} alice, bob", &mut users, &mut statuses);
-        assert_eq!(users, ["alice", "bob"]);
-    }
-
-    #[test]
-    fn seed_who_unrelated_system_message_is_noop() {
-        let mut users = Vec::new();
-        let mut statuses = HashMap::new();
-        seed_online_users_from_who("alice set status: away", &mut users, &mut statuses);
-        assert!(users.is_empty());
-        assert!(statuses.is_empty());
-    }
-
-    // ── parse_status_broadcast tests ─────────────────────────────────────────
-
-    #[test]
-    fn parse_status_set() {
-        let result = parse_status_broadcast("alice set status: busy");
-        assert_eq!(result, Some(("alice".to_owned(), "busy".to_owned())));
-    }
-
-    #[test]
-    fn parse_status_cleared() {
-        let result = parse_status_broadcast("alice cleared their status");
-        assert_eq!(result, Some(("alice".to_owned(), String::new())));
-    }
-
-    #[test]
-    fn parse_status_unrelated_message() {
-        assert!(parse_status_broadcast("alice joined").is_none());
-        assert!(parse_status_broadcast("hello world").is_none());
-    }
-
-    #[test]
-    fn parse_status_with_spaces_in_name() {
-        let result = parse_status_broadcast("my-agent set status: reviewing PR");
-        assert_eq!(
-            result,
-            Some(("my-agent".to_owned(), "reviewing PR".to_owned()))
-        );
-    }
-
-    // ── parse_kick_broadcast tests ──────────────────────────────────────────
-
-    #[test]
-    fn parse_kick_standard() {
-        let result = parse_kick_broadcast("alice kicked bob (token invalidated)");
-        assert_eq!(result, Some("bob"));
-    }
-
-    #[test]
-    fn parse_kick_hyphenated_names() {
-        let result = parse_kick_broadcast("room-host kicked my-agent (token invalidated)");
-        assert_eq!(result, Some("my-agent"));
-    }
-
-    #[test]
-    fn parse_kick_unrelated_message() {
-        assert!(parse_kick_broadcast("alice set status: busy").is_none());
-        assert!(parse_kick_broadcast("hello world").is_none());
-        assert!(parse_kick_broadcast("alice cleared their status").is_none());
-    }
-
-    #[test]
-    fn parse_kick_missing_suffix() {
-        assert!(parse_kick_broadcast("alice kicked bob").is_none());
-    }
-
-    #[test]
-    fn parse_kick_missing_target() {
-        assert!(parse_kick_broadcast("alice kicked  (token invalidated)").is_none());
-    }
-
-    // ── parse_subscription_broadcast tests ────────────────────────────────────
-
-    #[test]
-    fn parse_subscription_full() {
-        let result = parse_subscription_broadcast("alice subscribed to room-dev (tier: full)");
-        assert_eq!(result, Some(("alice".to_owned(), SubscriptionTier::Full)));
-    }
-
-    #[test]
-    fn parse_subscription_mentions_only() {
-        let result =
-            parse_subscription_broadcast("bob subscribed to my-room (tier: mentions_only)");
-        assert_eq!(
-            result,
-            Some(("bob".to_owned(), SubscriptionTier::MentionsOnly))
-        );
-    }
-
-    #[test]
-    fn parse_subscription_unsubscribed() {
-        let result = parse_subscription_broadcast("carol subscribed to test (tier: unsubscribed)");
-        assert_eq!(
-            result,
-            Some(("carol".to_owned(), SubscriptionTier::Unsubscribed))
-        );
-    }
-
-    #[test]
-    fn parse_subscription_unrelated_message() {
-        assert!(parse_subscription_broadcast("alice set status: busy").is_none());
-        assert!(parse_subscription_broadcast("hello world").is_none());
-        assert!(parse_subscription_broadcast("alice kicked bob (token invalidated)").is_none());
-    }
-
-    #[test]
-    fn parse_subscription_hyphenated_username() {
-        let result = parse_subscription_broadcast("my-agent subscribed to agent-room (tier: full)");
-        assert_eq!(
-            result,
-            Some(("my-agent".to_owned(), SubscriptionTier::Full))
-        );
-    }
-
-    // ── build_payload tests ───────────────────────────────────────────────────
-
-    #[test]
-    fn build_payload_plain_text_is_message_type() {
-        let payload = build_payload("hello world");
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        assert_eq!(v["type"], "message");
-        assert_eq!(v["content"], "hello world");
-    }
-
-    #[test]
-    fn build_payload_dm_command() {
-        let payload = build_payload("/dm alice hey there");
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        assert_eq!(v["type"], "dm");
-        assert_eq!(v["to"], "alice");
-        assert_eq!(v["content"], "hey there");
-    }
-
-    #[test]
-    fn build_payload_slash_command_becomes_command_type() {
-        let payload = build_payload("/kick alice");
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        assert_eq!(v["type"], "command");
-        assert_eq!(v["cmd"], "kick");
-    }
-
-    #[test]
-    fn build_payload_who_command() {
-        let payload = build_payload("/who");
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        assert_eq!(v["type"], "command");
-        assert_eq!(v["cmd"], "who");
-    }
-
-    #[test]
-    fn build_payload_dm_preserves_spaces_in_content() {
-        let payload = build_payload("/dm bob hello   world");
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        assert_eq!(v["content"], "hello   world");
-    }
-
-    // ── parse_dm_input tests ────────────────────────────────────────────────
-
-    #[test]
-    fn parse_dm_input_returns_dm_room_action() {
-        let action = parse_dm_input("/dm alice hello there").unwrap();
-        match action {
-            Action::DmRoom {
-                target_user,
-                content,
-            } => {
-                assert_eq!(target_user, "alice");
-                assert_eq!(content, "hello there");
-            }
-            _ => panic!("expected DmRoom action"),
-        }
-    }
-
-    #[test]
-    fn parse_dm_input_preserves_spaces_in_content() {
-        let action = parse_dm_input("/dm bob hello   world").unwrap();
-        match action {
-            Action::DmRoom { content, .. } => {
-                assert_eq!(content, "hello   world");
-            }
-            _ => panic!("expected DmRoom action"),
-        }
-    }
-
-    #[test]
-    fn parse_dm_input_returns_none_for_missing_content() {
-        assert!(parse_dm_input("/dm alice").is_none());
-    }
-
-    #[test]
-    fn parse_dm_input_returns_none_for_missing_user() {
-        assert!(parse_dm_input("/dm ").is_none());
-    }
-
-    #[test]
-    fn parse_dm_input_returns_none_for_non_dm() {
-        assert!(parse_dm_input("/who").is_none());
-        assert!(parse_dm_input("hello world").is_none());
-    }
-
-    #[test]
-    fn parse_dm_input_returns_none_for_user_only_with_trailing_space() {
-        // "/dm alice " has empty content after splitting
-        assert!(parse_dm_input("/dm alice ").is_none());
-    }
-
-    // ── wrap_input_display ──────────────────────────────────────────────────
-
-    #[test]
-    fn wrap_empty_string_returns_one_empty_row() {
-        assert_eq!(wrap_input_display("", 10), vec![""]);
-    }
-
-    #[test]
-    fn wrap_fits_on_one_line() {
-        assert_eq!(wrap_input_display("hello", 10), vec!["hello"]);
-    }
-
-    #[test]
-    fn wrap_exactly_at_boundary_stays_one_line() {
-        assert_eq!(wrap_input_display("abcde", 5), vec!["abcde"]);
-    }
-
-    #[test]
-    fn wrap_one_char_over_splits_to_two_rows() {
-        assert_eq!(wrap_input_display("abcdef", 5), vec!["abcde", "f"]);
-    }
-
-    #[test]
-    fn wrap_explicit_newline_splits_logical_lines() {
-        assert_eq!(
-            wrap_input_display("hello\nworld", 20),
-            vec!["hello", "world"]
-        );
-    }
-
-    #[test]
-    fn wrap_explicit_newline_at_end_gives_trailing_empty_row() {
-        assert_eq!(wrap_input_display("hi\n", 20), vec!["hi", ""]);
-    }
-
-    #[test]
-    fn wrap_combines_explicit_newlines_and_width_wrapping() {
-        assert_eq!(wrap_input_display("abcde\nfg", 3), vec!["abc", "de", "fg"]);
-    }
-
-    #[test]
-    fn wrap_width_zero_returns_lines_unsplit() {
-        assert_eq!(
-            wrap_input_display("a very long line", 0),
-            vec!["a very long line"]
-        );
-    }
-
-    #[test]
-    fn wrap_width_zero_with_newlines_splits_on_newlines_only() {
-        assert_eq!(wrap_input_display("foo\nbar", 0), vec!["foo", "bar"]);
-    }
-
-    #[test]
-    fn wrap_wide_chars_counted_by_display_width() {
-        let rows = wrap_input_display("\u{4e2d}\u{4e2d}\u{4e2d}", 4);
-        assert_eq!(rows, vec!["\u{4e2d}\u{4e2d}", "\u{4e2d}"]);
-    }
-
-    // ── cursor_display_pos ──────────────────────────────────────────────────
-
-    #[test]
-    fn cursor_at_start_of_empty_input() {
-        assert_eq!(cursor_display_pos("", 0, 10), (0, 0));
-    }
-
-    #[test]
-    fn cursor_at_end_of_single_line() {
-        assert_eq!(cursor_display_pos("hello", 5, 10), (0, 5));
-    }
-
-    #[test]
-    fn cursor_mid_single_line() {
-        assert_eq!(cursor_display_pos("hello", 2, 10), (0, 2));
-    }
-
-    #[test]
-    fn cursor_wraps_to_second_row() {
-        assert_eq!(cursor_display_pos("abcdef", 5, 5), (1, 0));
-    }
-
-    #[test]
-    fn cursor_after_explicit_newline() {
-        assert_eq!(cursor_display_pos("hi\n", 3, 20), (1, 0));
-    }
-
-    #[test]
-    fn cursor_on_second_explicit_line() {
-        assert_eq!(cursor_display_pos("foo\nbar", 7, 10), (1, 3));
-    }
-
-    #[test]
-    fn cursor_explicit_newline_combined_with_wrap() {
-        let s = "abc\ndefgh";
-        assert_eq!(cursor_display_pos(s, s.len(), 3), (2, 2));
-    }
-
-    #[test]
-    fn cursor_width_zero_no_wrapping() {
-        assert_eq!(cursor_display_pos("hello world", 5, 0), (0, 5));
-    }
-
-    #[test]
-    fn cursor_wide_char_advances_by_display_width() {
-        let s = "\u{4e2d}\u{4e2d}";
-        assert_eq!(cursor_display_pos(s, 3, 4), (0, 2));
-        assert_eq!(cursor_display_pos(s, 6, 4), (0, 4));
-    }
-
-    #[test]
-    fn cursor_wide_char_at_boundary_with_more_content() {
-        let s = "\u{4e2d}\u{4e2d}\u{4e2d}";
-        assert_eq!(cursor_display_pos(s, 6, 4), (1, 0));
-    }
-
-    #[test]
-    fn cursor_at_exact_line_boundary_no_more_content() {
-        assert_eq!(cursor_display_pos("abcde", 5, 5), (0, 5));
-    }
-
-    // ── apply_backslash_enter ───────────────────────────────────────────────
-
-    #[test]
-    fn backslash_enter_at_end_replaces_backslash_with_newline() {
-        let mut buf = String::from("hello\\");
-        let pos = apply_backslash_enter(&mut buf, 6);
-        assert_eq!(pos, Some(6));
-        assert_eq!(buf, "hello\n");
-    }
-
-    #[test]
-    fn backslash_enter_mid_buffer_replaces_at_cursor() {
-        let mut buf = String::from("foo\\bar");
-        let pos = apply_backslash_enter(&mut buf, 4);
-        assert_eq!(pos, Some(4));
-        assert_eq!(buf, "foo\nbar");
-    }
-
-    #[test]
-    fn backslash_enter_no_preceding_backslash_returns_none() {
-        let mut buf = String::from("hello");
-        let pos = apply_backslash_enter(&mut buf, 5);
-        assert_eq!(pos, None);
-        assert_eq!(buf, "hello");
-    }
-
-    #[test]
-    fn backslash_enter_cursor_at_start_returns_none() {
-        let mut buf = String::from("\\hello");
-        let pos = apply_backslash_enter(&mut buf, 0);
-        assert_eq!(pos, None);
-        assert_eq!(buf, "\\hello");
-    }
-
-    #[test]
-    fn backslash_enter_empty_buffer_returns_none() {
-        let mut buf = String::new();
-        let pos = apply_backslash_enter(&mut buf, 0);
-        assert_eq!(pos, None);
-    }
-
-    #[test]
-    fn backslash_enter_cursor_not_at_backslash_returns_none() {
-        let mut buf = String::from("a\\b");
-        let pos = apply_backslash_enter(&mut buf, 1);
-        assert_eq!(pos, None);
-        assert_eq!(buf, "a\\b");
-    }
-
-    #[test]
-    fn backslash_enter_double_backslash_replaces_last_one() {
-        let mut buf = String::from("foo\\\\");
-        let pos = apply_backslash_enter(&mut buf, 5);
-        assert_eq!(pos, Some(5));
-        assert_eq!(buf, "foo\\\n");
-    }
-
-    // ── find_at_context ───────────────────────────────────────────────────────
-
-    #[test]
-    fn find_at_context_no_at_returns_none() {
-        assert!(find_at_context("hello world", 11).is_none());
-    }
-
-    #[test]
-    fn find_at_context_bare_at_returns_at_with_empty_query() {
-        let input = "say @";
-        let result = find_at_context(input, input.len());
-        assert_eq!(result, Some((4, "")));
-    }
-
-    #[test]
-    fn find_at_context_partial_username() {
-        let input = "hello @ali";
-        let result = find_at_context(input, input.len());
-        assert_eq!(result, Some((6, "ali")));
-    }
-
-    #[test]
-    fn find_at_context_space_after_at_returns_none() {
-        let input = "@ alice";
-        assert!(find_at_context(input, input.len()).is_none());
-    }
-
-    #[test]
-    fn find_at_context_newline_in_query_returns_none() {
-        let input = "@ali\nce";
-        assert!(find_at_context(input, input.len()).is_none());
-    }
-
-    #[test]
-    fn find_at_context_cursor_before_at_returns_none() {
-        let input = "hello @ali";
-        assert!(find_at_context(input, 4).is_none());
-    }
-
-    #[test]
-    fn find_at_context_uses_last_at() {
-        let input = "@first @sec";
-        let result = find_at_context(input, input.len());
-        assert_eq!(result, Some((7, "sec")));
-    }
+    // Parse/display tests moved to parse.rs and display.rs.
 
     // ── handle_key unit tests ─────────────────────────────────────────────────
 
@@ -1565,137 +765,6 @@ mod tests {
         } else {
             panic!("expected Action::Send");
         }
-    }
-
-    // ── byte_offset_at_display_pos tests ─────────────────────────────────────
-
-    #[test]
-    fn byte_offset_single_row_exact_col() {
-        // "hello", width=80: row 0, col 3 → byte 3
-        assert_eq!(byte_offset_at_display_pos("hello", 0, 3, 80), 3);
-    }
-
-    #[test]
-    fn byte_offset_col_past_end_of_row_clamps_to_end() {
-        // "hi\nworld", row 0 only has 2 chars; col 10 → byte 2 (before '\n')
-        assert_eq!(byte_offset_at_display_pos("hi\nworld", 0, 10, 80), 2);
-    }
-
-    #[test]
-    fn byte_offset_second_logical_row() {
-        // "hello\nworld", row 1 col 3 → byte 9 ('l' in "world")
-        assert_eq!(byte_offset_at_display_pos("hello\nworld", 1, 3, 80), 9);
-    }
-
-    #[test]
-    fn byte_offset_soft_wrapped_row() {
-        // "abcdef" with width=3 wraps: row0="abc", row1="def"
-        // row 1, col 1 → byte 4 ('e')
-        assert_eq!(byte_offset_at_display_pos("abcdef", 1, 1, 3), 4);
-    }
-
-    #[test]
-    fn byte_offset_col_past_soft_wrapped_row_end() {
-        // "abcdef" width=3: row 0 ends at byte 3; col 10 clamps to byte 3
-        assert_eq!(byte_offset_at_display_pos("abcdef", 0, 10, 3), 3);
-    }
-
-    #[test]
-    fn byte_offset_end_of_input() {
-        // row beyond content → input.len()
-        assert_eq!(byte_offset_at_display_pos("hello", 5, 0, 80), 5);
-    }
-
-    // ── prev_word_start ───────────────────────────────────────────────────────
-
-    #[test]
-    fn prev_word_start_from_end_of_word() {
-        // "hello world", cursor at end → previous word start = 6 ("world")
-        assert_eq!(prev_word_start("hello world", 11), 6);
-    }
-
-    #[test]
-    fn prev_word_start_from_mid_word() {
-        // "hello world", cursor at 8 (mid "world") → word start = 6
-        assert_eq!(prev_word_start("hello world", 8), 6);
-    }
-
-    #[test]
-    fn prev_word_start_skips_trailing_whitespace() {
-        // "hello world  ", cursor at 13 (after trailing spaces) → "world" starts at 6
-        assert_eq!(prev_word_start("hello world  ", 13), 6);
-    }
-
-    #[test]
-    fn prev_word_start_at_beginning_returns_zero() {
-        assert_eq!(prev_word_start("hello", 0), 0);
-    }
-
-    #[test]
-    fn prev_word_start_from_first_word_returns_zero() {
-        // "hello world", cursor in "hello" → 0
-        assert_eq!(prev_word_start("hello world", 3), 0);
-    }
-
-    #[test]
-    fn prev_word_start_single_word_no_spaces() {
-        assert_eq!(prev_word_start("hello", 5), 0);
-    }
-
-    #[test]
-    fn prev_word_start_multiple_spaces_between_words() {
-        // "foo   bar", cursor at 9 → "bar" starts at 6
-        assert_eq!(prev_word_start("foo   bar", 9), 6);
-    }
-
-    #[test]
-    fn prev_word_start_unicode_word() {
-        // "α β", cursor at 5 (after β which is 2 bytes) → 3 (start of β)
-        let s = "α β"; // α=2 bytes, space=1, β=2 bytes → len=5
-        assert_eq!(prev_word_start(s, 5), 3);
-    }
-
-    // ── next_word_end ─────────────────────────────────────────────────────────
-
-    #[test]
-    fn next_word_end_from_start() {
-        // "hello world", cursor at 0 → end of "hello" = 5
-        assert_eq!(next_word_end("hello world", 0), 5);
-    }
-
-    #[test]
-    fn next_word_end_from_mid_word() {
-        // "hello world", cursor at 2 → end of "hello" = 5
-        assert_eq!(next_word_end("hello world", 2), 5);
-    }
-
-    #[test]
-    fn next_word_end_skips_leading_whitespace() {
-        // "hello world", cursor at 5 (space) → end of "world" = 11
-        assert_eq!(next_word_end("hello world", 5), 11);
-    }
-
-    #[test]
-    fn next_word_end_multiple_spaces() {
-        // "foo   bar", cursor at 3 → end of "bar" = 9
-        assert_eq!(next_word_end("foo   bar", 3), 9);
-    }
-
-    #[test]
-    fn next_word_end_at_end_returns_len() {
-        assert_eq!(next_word_end("hello", 5), 5);
-    }
-
-    #[test]
-    fn next_word_end_only_spaces_returns_len() {
-        assert_eq!(next_word_end("   ", 0), 3);
-    }
-
-    #[test]
-    fn next_word_end_unicode_word() {
-        // "α β", cursor at 0 → end of "α" = 2
-        let s = "α β";
-        assert_eq!(next_word_end(s, 0), 2);
     }
 
     // ── Alt+word-skip via handle_key ──────────────────────────────────────────
@@ -2405,43 +1474,6 @@ mod tests {
             state.mention.active,
             "mention picker should activate for Username params"
         );
-    }
-
-    // ── normalize_paste tests ───────────────────────────────────────────────
-
-    #[test]
-    fn normalize_paste_unix_newlines_unchanged() {
-        assert_eq!(normalize_paste("hello\nworld"), "hello\nworld");
-    }
-
-    #[test]
-    fn normalize_paste_windows_crlf_to_lf() {
-        assert_eq!(normalize_paste("hello\r\nworld"), "hello\nworld");
-    }
-
-    #[test]
-    fn normalize_paste_old_mac_cr_to_lf() {
-        assert_eq!(normalize_paste("hello\rworld"), "hello\nworld");
-    }
-
-    #[test]
-    fn normalize_paste_mixed_endings() {
-        assert_eq!(normalize_paste("a\r\nb\rc\nd"), "a\nb\nc\nd");
-    }
-
-    #[test]
-    fn normalize_paste_no_newlines() {
-        assert_eq!(normalize_paste("plain text"), "plain text");
-    }
-
-    #[test]
-    fn normalize_paste_empty() {
-        assert_eq!(normalize_paste(""), "");
-    }
-
-    #[test]
-    fn normalize_paste_multiple_crlf() {
-        assert_eq!(normalize_paste("a\r\n\r\nb"), "a\n\nb");
     }
 
     // ── Ctrl+D behavior (raw-mode TUI) ─────────────────────────────────────
